Added a simple implementation of a ShaderDatabase.
Basically, requests are done to the shader database to load them from the database and use them as shaders. If they aren't found in the database, or if the shader's file is more recent that what is in the shader database, then the entry in the database is removed and a request is done to the shader compiler. When the shader compiler returns a compiled shader, an entry is added to the database.

To improve:
* The ShaderCompiler shouldn't have the information about whether or not a shader file was modified, only the shader watcher should.
* The ShaderCompiler should only have to compile a single shader key, instead of a whole shader family.